requirements.txt
streamlit>=1.30.0
tensorflow>=2.12.0
opencv-python-headless>=4.8.0
numpy>=1.23.0
pillow>=9.5.0
pandas>=1.5.0